+++ src/pages/Dashboard/Settings/index.tsx
@@ -11,7 +11,7 @@ import UpdateMiniHub from './UpdateMiniHub';
 import { useNavigate } from 'react-router-dom';
 import ShareConnections from './ShareConnections';
 import { peers } from '../../../lib';
-import Toggle from '../../../components/UI/Toggle';
+import Folders from './Folders';
 
 export function Settings() {
   const navigate = useNavigate();
@@ -22,8 +22,8 @@ export function Settings() {
     showAddConnections,
     setShowAddConnections,
     setShowOnboard,
-    folderStatus,
-    toggleFolderStatus,
+    showFoldersTheme,
+    setShowFoldersTheme,
   } = useContext(appContext);
   const [showShutdown, setShutdown] = useState(false);
   const [showWallpaper, setShowWallpaper] = useState(false);
@@ -66,6 +66,7 @@ export function Settings() {
       <DesktopConnect display={showDesktopConnect} dismiss={() => setShowDesktopConnect(false)} />
       <ShareConnections display={showShareConnections} dismiss={() => setShowShareConnections(false)} />
       <BatteryOptimisation display={showBatteryOptimisation} dismiss={() => setShowBatteryOptimisation(false)} />
+      <Folders display={showFoldersTheme} dismiss={() => setShowFoldersTheme(false)} />
       <div className="flex flex-col h-full min-h-[700px]">
         <InScreenTitleBar onExit={dismiss} />
         <div className="pt-4 px-4 pb-4 flex flex-col h-full max-w-xl mx-auto w-full">
@@ -174,12 +175,20 @@ export function Settings() {
                 </svg>
               </div>
             </div>
-            <div className="relative core-black-contrast-2 py-4 px-5 rounded cursor-pointer">
-              Folder layout
+            
+            
+            <div onClick={() => setShowFoldersTheme(true)} className="relative core-black-contrast-2 py-4 px-5 rounded cursor-pointer">
+              Folders
               <div className="absolute right-0 top-0 h-full px-5 flex items-center">
-                <Toggle checkedStatus={folderStatus} onChange={toggleFolderStatus} />
+                <svg width="8" height="12" viewBox="0 0 8 12" fill="none" xmlns="http://www.w3.org/2000/svg">
+                  <path
+                    d="M7.04984 5.99995L1.37504 11.6501L0.500244 10.7501L5.24984 5.99995L0.500244 1.24975L1.40024 0.349747L7.04984 5.99995Z"
+                    fill="#F4F4F5"
+                  />
+                </svg>
               </div>
             </div>
+
             <div
               onClick={() => setShutdown(true)}
               className="relative text-status-red core-black-contrast py-4 px-5 rounded cursor-pointer"